+++ src/main/java/com/buuz135/industrial/utils/apihandlers/crafttweaker/CTBioReactor.java
@@ -12,7 +12,6 @@ import stanhebben.zenscript.annotations.ZenMethod;
 @ZenClass("mods.industrialforegoing.BioReactor")
 public class CTBioReactor {
 
-
     @ZenMethod
     public static void add(IItemStack input) {
         BioReactorEntry entry = new BioReactorEntry((ItemStack) input.getInternal());